Add scrollbar on connections dropdown menu (#992)
Why?

When you have a lot of connections registered in your localStorage, you will reach a limit on the connections dropdown sizing. For now, you cannot scroll on it and you must zoom out to see the hidden items. This PR fix this behaviour by adding a scrollbar when dropdown menu content exceed the 98% of the current browser size (we need a small margin because of the native styling of the Bootstrap component)
Boyscout

    Fix Cypress test process ID generation to make it unique